Move failing unit test into ticket-specific case
